progress toward integrating with frontend
