+++ src/frontend/index.html
@@ -14,7 +14,6 @@
     <script src="https://unpkg.com/@babel/standalone/babel.min.js"></script>
     <script src="https://unpkg.com/xstate@4/dist/xstate.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/@xstate/react/dist/xstate-react.umd.min.js"></script>
-
     <link rel="stylesheet" href="/spinners.css" type="text/css" />
 
     <script>
