+++ src/proto.py
@@ -3,6 +3,7 @@ Main web application service. Serves the static frontend as well as
 API routes for transcription, language model generation and text-to-speech.
 """
 
+from os import pipe
 from pathlib import Path
 from string import punctuation
 import modal
@@ -13,48 +14,31 @@ from .llm_zephyr import Zephyr
 
 from .common_proto import app
 
-# constants
-# LLM_TPS = 60
+pipeline_start_time = time.time()
+def timeprint(*args):
+    print(f"{time.time() - pipeline_start_time}s", *args)
 
-# @app.cls(
-#     container_idle_timeout=300,
-#     timeout = 180,
-#     concurrency_limit=5,
-# )
-# class LLM:
-#     @modal.enter()
-#     def load_model(self):
-#         print("Loading LLM model")
-#         time.sleep(3)
-#         print("LLM model loaded")
+static_path = Path(__file__).with_name("frontend").resolve()
 
-#     @modal.method()
-#     def prewarm(self):
-#         # no-op to prewarm LLM model instance
-#         pass
-
-#     @modal.method(is_generator=True)
-#     def generate(self, text):
-#         # text is the output of whisper
-        
-#         # simulate llm with response streaming
-#         for word in text.split(" "):
-#             # Why is the final word an empty string here?
-#             print(f"Yielding word from llm.generate.remote_gen: |{word}|")
-#             time.sleep(1/LLM_TPS)
-#             yield word
-
-@app.function( 
+@app.function(
+    # mounts=[modal.Mount.from_local_dir(static_path, remote_path="/assets")],
     container_idle_timeout=300,
-    timeout = 180,
+    timeout=600,
 )
 @modal.asgi_app()
 def web():
     from fastapi import FastAPI, Request, Response, WebSocket
     from fastapi.responses import Response, StreamingResponse
     from fastapi.staticfiles import StaticFiles
-
+    import numpy as np
+    import wave
+    import io
+    
     web_app = FastAPI()
+       
+    # serve static files
+    # web_app.mount("/", StaticFiles(directory="/assets", html=True))
+
     whisper = Whisper()
     zephyr = Zephyr()
     xtts = XTTS()
@@ -70,35 +54,100 @@ def web():
             i.get()
 
         return Response(status_code=200)
+    
+    def float32_to_wav(float32_buffer, sample_rate=44100):
+        # Convert to int16
+        int16_data = (float32_buffer * 32767).astype(np.int16)
+        
+        # Create an in-memory binary stream
+        byte_io = io.BytesIO()
+        
+        # Create WAV file in the binary stream
+        with wave.open(byte_io, 'wb') as wav_file:
+            wav_file.setnchannels(1)  # Mono
+            wav_file.setsampwidth(2)  # 2 bytes per sample (16 bits)
+            wav_file.setframerate(sample_rate)
+            wav_file.writeframes(int16_data.tobytes())
+        
+        # Get the WAV data as bytes
+        wav_data = byte_io.getvalue()
+        return wav_data
 
     @web_app.websocket("/pipeline")
     async def websocket_endpoint(websocket: WebSocket):
         await websocket.accept()
 
+        # temp: testing pipeline timing
+        global pipeline_start_time
+        pipeline_start_time= time.time()
+
+
+        # TEMP: make very simple
+        buffer = b""
+        try:
+            while True:
+                data = await websocket.receive_bytes()
+                
+                if data == b'<END>':
+                    # Handle end of transmission
+                    if buffer:
+                        # Process any remaining data in the buffer
+                        float32_buffer = np.frombuffer(buffer, dtype=np.float32)
+                        wav_data = float32_to_wav(float32_buffer)
+                        await websocket.send_bytes(wav_data)
+                    break
+                
+                buffer += data
+                
+                # Process as many complete float32 samples as possible
+                num_complete_samples = len(buffer) // 4  # 4 bytes per float32
+                if num_complete_samples > 0:
+                    complete_data = buffer[:num_complete_samples * 4]
+                    buffer = buffer[num_complete_samples * 4:]
+                    
+                    float32_buffer = np.frombuffer(complete_data, dtype=np.float32)
+                    wav_data = float32_to_wav(float32_buffer)
+                    
+                    await websocket.send_bytes(wav_data)
+        
+        except Exception as e:
+            print(f"Error: {e}")
+        finally:
+            await websocket.close()
+        return
+
         # Step 1: User streams their input in via WebSocket
         async def user_input_stream_gen():
-            print("Receiving user WAV input stream...")
+            i = 0
             while True:
+                timeprint("websocket.receive_bytes waiting for WAV chunk", i)
                 wav_bytes = await websocket.receive_bytes()
                 if wav_bytes == b"<END>":
-                    print("Received end of stream")
+                    timeprint("websocket.receive_bytes received <END> signal")
                     break
-                print("Received WAV chunk")
+                timeprint("websocket.receive_bytes received WAV chunk", i)
+                i += 1
                 yield wav_bytes
 
         # We parallel transcribe user input chunks
         transcribe_futures = []
+        i = 0
         async for chunk in user_input_stream_gen():
+            timeprint("TRANSCRIPTION: starting WAV chunk", i)
+            i += 1
             # the moment the chunk arrives, we start transcribing it
             transcribe_futures.append(whisper.transcribe.spawn(chunk))
         
         # We await all transcription chunks
         transcript_chunks = []
-        for i in transcribe_futures:
-            transcript_chunk = i.get()
+        i = 0
+        for id in transcribe_futures:
+            timeprint("TRANSCRIPTION: awaiting chunk", i)
+            i += 1
+            transcript_chunk = id.get()
             transcript_chunks.append(transcript_chunk)
         transcript = " ".join(transcript_chunks)
-        print("Transcript Complete: ", transcript)
+        timeprint("TRANSCRIPTION: complete", transcript)
 
         # We now have a single transcript to send to the LLM
         llm_response_stream_gen = zephyr.generate.remote_gen(transcript)
@@ -106,18 +155,39 @@ def web():
         # llm_response_stream_gen will yield words, which we'll want to 
         # accumulate together into sentences for more natural-sounding TTS.
         punctuation = [".", "?", "!", ":", ";", "*"]
+        max_words = 5
         def tts_input_stream_acccumulator(text_stream):
+            chunk_i = 0
             current_chunk = ""
             for word in text_stream:
+                # receives yields from LLM
+                timeprint("LLM GENERATION: yielded word", word)
+
+                # TODO: explore why so many empty words are generated
+
                 current_chunk += word + " "
+
+                # yield if we're above max words
+                if len(current_chunk.split(" ")) > max_words:
+                    # yields sentences to TTS
+                    timeprint(f"TTS: starting sentence {chunk_i}:", current_chunk)
+                    chunk_i += 1
+                    yield current_chunk
+                    current_chunk = ""
+                    continue
+
+                # yield if we're at punctuation
                 for p in punctuation:
                     if p in word:
-                        print("Sentence from LLM ready for TTS: ", current_chunk)
+                        # yields sentences to TTS
+                        timeprint(f"TTS: starting sentence {chunk_i}:", current_chunk)
+                        chunk_i += 1
                         yield current_chunk
                         current_chunk = ""
                         break
             # last chunk
             if current_chunk != "":
+                timeprint(f"TTS: last sentence {chunk_i}:", current_chunk)
                 yield current_chunk
         
         # Stream the word-by-word response from the LLM into the accumulator
@@ -125,82 +195,24 @@ def web():
 
         # Stream the sentences from the accumulator into the TTS service
         tts_output_stream_gen = xtts.speak.map(tts_input_stream_gen)
+        i = 0
         async for wavBytesIO in tts_output_stream_gen:
             # Stream the WAV bytes from the TTS service back to the client
-            print("TTS sentence completed")
+            timeprint(f"TTS: sentence {i} completed, sending to client")
             await websocket.send_bytes(wavBytesIO.getvalue())
+            timeprint(f"TTS: sentence {i} sent to client")
+            i += 1
 
-        print("Pipeline completed, all bytes sent, closing websocket")
+        timeprint("PIPELINE COMPLETE")
         await websocket.close()
+        
     return web_app
 
 
-@app.local_entrypoint()
-def main():
-    print("hello!")
-    # import sounddevice as sd
-    # import numpy as np
-    # from scipy.io import wavfile
 
-    # print("Starting dev script")
-    # user_input = user_input_generator()
-    # whisper = Whisper()
-    # llm = LLM()
-    # tts = XTTS()
-
-    # # ensure warm instances of everything
-    # prewarm_ids = []
-    # prewarm_ids.append(whisper.prewarm.spawn())
-    # prewarm_ids.append(llm.prewarm.spawn())
-    # prewarm_ids.append(tts.prewarm.spawn())
-    # for i in prewarm_ids:
-    #     i.get()
-
-    # # parallel whisper chunks
-    # transcipt_chunks = whisper.transcribe.map(user_input)
-    # transcript = " ".join(list(transcipt_chunks)) # await all
-
-    # global user_finish_time
-    # print(f"Time since end of user speech to TRANSCRIPTION: {time.time() - user_finish_time}s")
-
-    # # single llm response, streamed back
-    # llm_response_stream = llm.generate.remote_gen(transcript)
-
-    # # accumulate text stream into larger chunks split by punctuation, since TTS needs slightly longer chunks to sound natural
-    # punctuation = [".", "?", "!", ":", ";", "*"]
-    # def chunk_generator(text_stream):
-    #     current_chunk = ""
-    #     i = 0
-    #     for word in text_stream:
-    #         if i == 0:
-    #             print(f"Time since end of user speech to FIRST LLM TOKEN: {time.time() - user_finish_time}s")
-    #         i += 1
-    #         current_chunk += word + " "
-    #         for p in punctuation:
-    #             if p in word:
-    #                 yield current_chunk
-    #                 current_chunk = ""
-    #                 break
-        
-    #     # last chunk
-    #     if current_chunk != "":
-    #         yield current_chunk
-
-    # chunk_input = chunk_generator(llm_response_stream)
-    # tts_chunks = tts.speak.map(chunk_input)
-
-    # i = 0
-    # for chunk in tts_chunks:
-    #     if i == 0:
-    #         print(f"Time since end of user speech to FIRST TTS CHUNK: {time.time() - user_finish_time}s")
-
-    #     # play WAV
-    #     sample_rate, data = wavfile.read(chunk)
-    #     sd.play(data, sample_rate)
-    #     sd.wait()
-
-    #     with open(f"output_{i}.wav", "wb") as f:
-    #         f.write(chunk.getvalue())
-    
-    #     i += 1
+## TODO: 
+# current pipeline takes 8.4sec from end of user speech to first LLM token
+# bottlenecks
+#   - LLM response generation takes 2sec to output the first sentence.
+#   - TTS 2.5sec on that first sentence
